Remove obsolete UserData.set_matchmaking_parameters
